+++ entities/video.go
@@ -2,17 +2,17 @@ package entities
 
 type Video struct {
 	Type              string `json:"@type"`
-	Duration          int    `json:"duration"`
-	Width             int    `json:"width"`
-	Height            int    `json:"height"`
+	Duration          int64  `json:"duration"`
+	Width             int64  `json:"width"`
+	Height            int64  `json:"height"`
 	FileName          string `json:"file_name"`
 	MimeType          string `json:"mime_type"`
 	HasStickers       bool   `json:"has_stickers"`
 	SupportsStreaming bool   `json:"supports_streaming"`
 	Minithumbnail     struct {
 		Type   string `json:"@type"`
-		Width  int    `json:"width"`
-		Height int    `json:"height"`
+		Width  int64  `json:"width"`
+		Height int64  `json:"height"`
 		Data   string `json:"data"`
 	} `json:"minithumbnail"`
 	Thumbnail struct {
@@ -20,13 +20,13 @@ type Video struct {
 		Format struct {
 			Type string `json:"@type"`
 		} `json:"format"`
-		Width  int `json:"width"`
-		Height int `json:"height"`
+		Width  int64 `json:"width"`
+		Height int64 `json:"height"`
 		File   struct {
 			Type         string `json:"@type"`
-			Id           int    `json:"id"`
-			Size         int    `json:"size"`
-			ExpectedSize int    `json:"expected_size"`
+			Id           int64  `json:"id"`
+			Size         int64  `json:"size"`
+			ExpectedSize int64  `json:"expected_size"`
 			Local        struct {
 				Type                   string `json:"@type"`
 				Path                   string `json:"path"`
@@ -34,9 +34,9 @@ type Video struct {
 				CanBeDeleted           bool   `json:"can_be_deleted"`
 				IsDownloadingActive    bool   `json:"is_downloading_active"`
 				IsDownloadingCompleted bool   `json:"is_downloading_completed"`
-				DownloadOffset         int    `json:"download_offset"`
-				DownloadedPrefixSize   int    `json:"downloaded_prefix_size"`
-				DownloadedSize         int    `json:"downloaded_size"`
+				DownloadOffset         int64  `json:"download_offset"`
+				DownloadedPrefixSize   int64  `json:"downloaded_prefix_size"`
+				DownloadedSize         int64  `json:"downloaded_size"`
 			} `json:"local"`
 			Remote struct {
 				Type                 string `json:"@type"`
@@ -44,15 +44,15 @@ type Video struct {
 				UniqueId             string `json:"unique_id"`
 				IsUploadingActive    bool   `json:"is_uploading_active"`
 				IsUploadingCompleted bool   `json:"is_uploading_completed"`
-				UploadedSize         int    `json:"uploaded_size"`
+				UploadedSize         int64  `json:"uploaded_size"`
 			} `json:"remote"`
 		} `json:"file"`
 	} `json:"thumbnail"`
 	Video struct {
 		Type         string `json:"@type"`
-		Id           int    `json:"id"`
-		Size         int    `json:"size"`
-		ExpectedSize int    `json:"expected_size"`
+		Id           int64  `json:"id"`
+		Size         int64  `json:"size"`
+		ExpectedSize int64  `json:"expected_size"`
 		Local        struct {
 			Type                   string `json:"@type"`
 			Path                   string `json:"path"`
@@ -60,9 +60,9 @@ type Video struct {
 			CanBeDeleted           bool   `json:"can_be_deleted"`
 			IsDownloadingActive    bool   `json:"is_downloading_active"`
 			IsDownloadingCompleted bool   `json:"is_downloading_completed"`
-			DownloadOffset         int    `json:"download_offset"`
-			DownloadedPrefixSize   int    `json:"downloaded_prefix_size"`
-			DownloadedSize         int    `json:"downloaded_size"`
+			DownloadOffset         int64  `json:"download_offset"`
+			DownloadedPrefixSize   int64  `json:"downloaded_prefix_size"`
+			DownloadedSize         int64  `json:"downloaded_size"`
 		} `json:"local"`
 		Remote struct {
 			Type                 string `json:"@type"`
@@ -70,7 +70,7 @@ type Video struct {
 			UniqueId             string `json:"unique_id"`
 			IsUploadingActive    bool   `json:"is_uploading_active"`
 			IsUploadingCompleted bool   `json:"is_uploading_completed"`
-			UploadedSize         int    `json:"uploaded_size"`
+			UploadedSize         int64  `json:"uploaded_size"`
 		} `json:"remote"`
 	} `json:"video"`
 }
