+++ incomingevents/file.go
@@ -1,9 +1,9 @@
 package incomingevents
 
 type DownloadFileResponse struct {
-	Id           int32 `json:"id"`
-	Size         int   `json:"size"`
-	ExpectedSize int   `json:"expected_size"`
+	Id           int64 `json:"id"`
+	Size         int64 `json:"size"`
+	ExpectedSize int64 `json:"expected_size"`
 	Local        struct {
 		Type                   string `json:"@type"`
 		Path                   string `json:"path"`
@@ -11,8 +11,8 @@ type DownloadFileResponse struct {
 		CanBeDeleted           bool   `json:"can_be_deleted"`
 		IsDownloadingActive    bool   `json:"is_downloading_active"`
 		IsDownloadingCompleted bool   `json:"is_downloading_completed"`
-		DownloadOffset         int    `json:"download_offset"`
-		DownloadedPrefixSize   int    `json:"downloaded_prefix_size"`
+		DownloadOffset         int64  `json:"download_offset"`
+		DownloadedPrefixSize   int64  `json:"downloaded_prefix_size"`
 		DownloadedSize         int64  `json:"downloaded_size"`
 	} `json:"local"`
 	Remote struct {
@@ -21,6 +21,6 @@ type DownloadFileResponse struct {
 		UniqueId             string `json:"unique_id"`
 		IsUploadingActive    bool   `json:"is_uploading_active"`
 		IsUploadingCompleted bool   `json:"is_uploading_completed"`
-		UploadedSize         int    `json:"uploaded_size"`
+		UploadedSize         int64  `json:"uploaded_size"`
 	} `json:"remote"`
 }
